Move createStrictSelect to data/utils and add support for preserving name.
