+++ assets/js/googlesitekit/data/utils.js
@@ -20,6 +20,7 @@
  * External dependencies
  */
 import invariant from 'invariant';
+import mapValues from 'lodash/mapValues';
 
 /**
  * WordPress dependencies
@@ -281,4 +282,37 @@ export const commonStore = {
 	actions: commonActions,
 	controls: commonControls,
 	reducer: passthroughReducer,
+};
+
+/**
+ * Creates a strict version of registry.select for ensuring that a selector is resolved at the time of calling.
+ *
+ * Not intended to be used directly. This is useful in the context of validation functions
+ * to save checking for undefined on every result.
+ *
+ * Given the registry.select function instance, a new function is returned
+ * with the same API as `select()` but will throw an error if the result
+ * of the selector function is `undefined`.
+ *
+ * @since n.e.x.t
+ * @private
+ *
+ * @param {Function} select The registry.select function.
+ * @return {Function} The strict version of registry.select.
+ */
+export const createStrictSelect = ( select ) => ( storeName ) => {
+	return mapValues(
+		select( storeName ),
+		( selector, selectorName ) => {
+			const strictSelector = ( ...args ) => {
+				const returnValue = selector( ...args );
+				invariant( returnValue !== undefined, `${ selectorName }(...) is not resolved` );
+				return returnValue;
+			};
+			// Preserve the function's name by setting it on the new function.
+			// Function.name is read-only so we must use Object.defineProperty.
+			Object.defineProperty( strictSelector, 'name', { value: selectorName } );
+			return strictSelector;
+		}
+	);
 };
